Remove old cclw sites folder and move all across to new theme
