--- src/sites/cclw/styles/styles.scss
@@ -1,88 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-// Import base styles before overriding
-// @import "../../../styles/main.scss";
-
-@mixin cclw-dark-gradient() {
-  background: linear-gradient(134.5deg, rgba(43, 47, 73, 1) 0%, rgba(43, 47, 73, 1) 33%, rgba(21, 116, 209, 1) 66%, rgba(21, 116, 209, 1) 100%);
-  background-repeat: no-repeat;
-}
-
-@font-face {
-  font-family: "LL Brown";
-  src: url("/fonts/LL-Brown/lineto-brown-bold.woff2") format("woff2"), url("/fonts/LL-Brown/lineto-brown-bold.woff") format("woff");
-  /* font-display: swap; */
-  font-weight: 500;
-}
-
-#cclw h2 {
-  @apply font-normal;
-}
-
-#cclw .cclw-font {
-  font-family: "LL Brown";
-}
-
-#cclw .gradient-container {
-  @include cclw-dark-gradient;
-}
-
-#cclw header {
-  border-color: rgba(255, 255, 255, 0.5);
-}
-
-#cclw nav {
-  a {
-    @apply h-full block py-2 px-4;
-    &:hover,
-    &.active {
-      @apply bg-blue-400;
-    }
-  }
-}
-
-#cclw footer {
-  @apply bg-cclw-dark text-white;
-
-  h5 {
-    @apply text-white;
-  }
-
-  a {
-    @apply text-white underline;
-    &:hover {
-      @apply text-blue-200;
-    }
-  }
-
-  .footer__section {
-    @apply text-sm;
-    > * {
-      @apply mb-4;
-    }
-    h5 {
-      @apply mb-4;
-    }
-  }
-
-  .footer__social-links {
-    a {
-      display: block;
-      width: 20px;
-      height: 20px;
-    }
-  }
-
-  .footer__base {
-    background-image: url("/images/cclw/angles_2.svg");
-    background-position: top;
-    background-repeat: repeat-x;
-    @apply text-cclw-dark w-full h-[232px] flex items-end overflow-hidden lg:h-[184px];
-
-    a {
-      @apply text-cclw-dark;
-    }
-  }
-}
